Updated add and removeFavorite functions to be in happy hour context, updated fetch calls and set functions to setHappyHours after button is clicked. Happy hour now successfully adds or removes from database on click and renders appropriate button by a ternary on DOM.
