+++ src/components/happyhour/HappyHourList.js
@@ -1,18 +1,16 @@
 import React, { useContext, useEffect, useState } from "react"
 import { HappyHourContext } from "./HappyHourProvider.js"
 import { useHistory, useLocation } from 'react-router-dom'
-import CardDeck from 'react-bootstrap/CardDeck'
-import { FavoriteContext } from "../favorites/FavoriteProvider.js"
 import { HappyHourCard } from "./HappyHourCard.js"
 import "./HappyHour.css"
 import Row from 'react-bootstrap/Row'
 import Container from 'react-bootstrap/Container'
+import { FavoriteContext } from "../favorites/FavoriteProvider.js"
 
 export const HappyHourList = () => {
 
     const {happyhours, getHappyHours, searchTerms, getHappyHourSearch} = useContext(HappyHourContext)
-    const {addFavorite} = useContext(FavoriteContext)
-    const [active, setActive] = useState(false)
+    
 
     const history = useHistory()
 
@@ -32,19 +30,19 @@ export const HappyHourList = () => {
         <>
             <h1 className="happy__hours_title">{weekday ? `${weekday}'s Happy Hours` : "Today's Happy Hours"}</h1>
             <div className="happy_hours">
-            <Container>
-                <Row className="justify-content-md-left">
+                <Container>
+                    <Row className="justify-content-md-left">
            
-                {
+                    {
                     happyhours.map(happyhour => {
                         
                         return <HappyHourCard key={happyhour.id} happyhour={happyhour} />
 
-                    })
-                }
+                        })
+                    }
             
-                </Row>
-            </Container>
+                    </Row>
+                </Container>
             </div>
         </>
     )
