+++ src/components/happyhour/HappyHourProvider.js
@@ -39,8 +39,33 @@ export const HappyHourProvider = (props) => {
             .then(setHappyHours)
     }
 
+    const addFavorite = (favorite) => {
+        
+        return fetch("http://localhost:8000/favorites", {
+            method: "POST",    
+            headers:{
+                "Authorization": `Token ${localStorage.getItem("hhs_token")}`,
+                "Content-Type": "application/json"
+            },
+            body: JSON.stringify({
+                "happy_hour": favorite
+            })
+        })
+       
+    }
+
+    const removeFavorite = happyhourId => {
+        return fetch(`http://localhost:8000/favorites/${happyhourId}`, {
+            method: "DELETE",
+            headers:{
+                "Authorization": `Token ${localStorage.getItem("hhs_token")}`
+            }
+        })
+            
+    }
+
     return (
-        <HappyHourContext.Provider value={{ happyhours, searchTerms, getHappyHours, getHappyHourSearch, setTerms }}>
+        <HappyHourContext.Provider value={{ happyhours, searchTerms, getHappyHours, getHappyHourSearch, setTerms, addFavorite, removeFavorite }}>
             {props.children}
         </HappyHourContext.Provider>
     )
